refactor(wallet): fix circular import of hardware response code type
